requirements.txt
discord.py[voice]
yt-dlp
python-dotenv
PyNaCl
gTTS
